ml_models/disease_vision/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow
scikit-learn
tqdm
matplotlib
numpy
